refactor: split single-keyword hook types
