+++ src/CADShared/Cad/Editor/Input/SingleKeyWordHook.cs
@@ -236,42 +236,3 @@ public sealed class SingleKeyWordHook : IDisposable
 
     #endregion
 }
-/// <summary>
-/// 单文本关键字钩子扩展
-/// </summary>
-public static class SingleKeywordHookEx
-{
-    /// <summary>
-    /// 钩取单文本关键字
-    /// </summary>
-    /// <param name="keywords">关键字集合</param>
-    /// <param name="workType">工作模式</param>
-    /// <returns>单文本关键字类(需要using)</returns>
-    public static SingleKeyWordHook HookSingleKeyword(this KeywordCollection keywords, SingleKeyWordWorkType workType = SingleKeyWordWorkType.WRITE_LINE)
-    {
-        var singleKeyWordHook = new SingleKeyWordHook(workType);
-        singleKeyWordHook.AddKeys(keywords);
-        return singleKeyWordHook;
-    }
-}
-
-/// <summary>
-/// 单关键字工作模式
-/// </summary>
-public enum SingleKeyWordWorkType : byte
-{
-    /// <summary>
-    /// Esc模式
-    /// </summary>
-    ESCAPE,
-    
-    /// <summary>
-    /// Enter模式
-    /// </summary>
-    ENTER,
-    
-    /// <summary>
-    /// Write Line 模式
-    /// </summary>
-    WRITE_LINE,
-}
