+++ src/CADShared/Cad/Editor/Input/SingleKeyWordWorkType.cs
@@ -1,0 +1,22 @@
+namespace Fs.Fox.Cad;
+
+/// <summary>
+/// 单关键字工作模式
+/// </summary>
+public enum SingleKeyWordWorkType : byte
+{
+    /// <summary>
+    /// Esc模式
+    /// </summary>
+    ESCAPE,
+
+    /// <summary>
+    /// Enter模式
+    /// </summary>
+    ENTER,
+
+    /// <summary>
+    /// Write Line 模式
+    /// </summary>
+    WRITE_LINE,
+}
